move existing resources into new structure folder
